[many features] review later 1. new g_CreateTilesFromChains_ which creates tiles_str and tiles_end exactly starting and ending at the beginning and end of the chain 2. reform_tiles and interface of functions are modified to adapted to the new g_CreateTilesFromChains_ 3. disable trimTiles in the mapTilesFromAnchors, otherwise the tiles_str and tiles_end created by g_CreateTilesFromChains_ will be modified again,  not tested [todo]

+++ src/cords.cpp
@@ -274,8 +274,8 @@ void print_cords(String<uint64_t> & cords, CharString header)
                   << get_cord_strand(cords[i]) << " "
                   << get_cord_x(cords[i])  << " "
                   << get_cord_y (cords[i]) << " " 
-                  << get_cord_x(cords[i]) - prex << " "
-                  << get_cord_y(cords[i]) - prey << " "
+                  << int64_t(get_cord_x(cords[i]) - prex) << " "
+                  << int64_t(get_cord_y(cords[i]) - prey) << " "
                   << length(cords) << "\n";
         prex = get_cord_x(cords[i]);
         prey = get_cord_y(cords[i]);
